small patchup

+++ Sonic12Decomp/Audio.hpp
@@ -16,7 +16,7 @@ struct TrackInfo {
 struct MusicPlaybackInfo {
 #if RETRO_USING_SDL
     OggVorbis_File vorbisFile;
-    uint audioLen;
+    unsigned long long audioLen;
     int vorbBitstream;
     SDL_AudioSpec spec;
     SDL_AudioStream *stream;
